[DEV-4794] extracting helper fn "getUniqueAncestorPaths"
extracting duplicate logic and adding tests. This fn
is responsible for getting the unique ancestors of
all checked nodes so that we make an api call for the
checked nodes exactly once.

this logic is pretty much tested twice. Here and also
in the didMount logic for each checkbox tree where
we ensure each ancestor is only requested one time.
